Allow the user to use nice names when dealing with servers and sites.
The concept behind this change is simple: nobody remembers server IDs and often times when doing some tasks, it is far easier to just use the name. This change lets you do just that.

Since all of the existing commands used the same naming convention for server and site, it was easy enough to find-and-replace all of those existing instances and replace them with new functions on the BaseCommand which can do a lookup. Originally, I had considered wrapping the input to allow for it to "just work" but given that the input happens outside of the command context, there was not a simple way to get the servers. It needed to be done at the command level.

The servers and sites are cached using the existing config file.

A new `site:cache-clear` command was added to remove the cache in case of a problem.

This could likely be improved dramatically, but for now, this is a great start.